v0.1.2-feat: Introduced equivariant policy and invariant critic
- Updated imports in `__init__.py` to use `SymmOnPolicyRunner`.
- Removed `actor_critic_symm_equivariant_nn.py` as it is no longer needed.
- Updated `play_symm.py` and `train_symm.py` to use `SymmOnPolicyRunner`.
- Enhanced `EquivEmpiricalNormalization` in `normalizer.py` to use new statistics functions.
- Updated `pyproject.toml` to require a specific version of `symm-learning`.
